% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacf.R
\name{cacf}
\alias{cacf}
\title{Complex Autocorrelation Function Estimation}
\usage{
cacf(x, lag.max = NULL, type = c("correlation", "covariance"), plot = TRUE)
}
\arguments{
\item{x}{vector of complex variables.}

\item{lag.max}{maximum number of lags. See \link[stats](acf) for more details.}

\item{type}{character string giving the type of cacf to be computed. Allowed values
are "correlation" (the default) and "covariance". Will be partially matched.}

\item{plot}{logical. If \code{TRUE} (the default) the cacf is plotted on complex plane
and as two linear graphs for real and imaginary parts.}
}
\value{
An object of class "cacf", which is a list with the following elements:
\itemize{
\item \code{lag} A three dimensional array containing the lags at which the cacf is estimated.
\item \code{acf} An array with the same dimensions as lag containing the estimated cacf.
\item \code{type} The type of correlation (same as the type argument).
\item \code{n.used} The number of observations in the time series.
\item \code{series} The name of the series x.
}

#' @references \itemize{
\item Svetunkov, S. (2022) Complex Autoregressions. In Press.
}
}
\description{
The function computes (and by default plots) estimates of the complex autocovariance
or complex autocorrelation function.
}
\details{
For \code{type="correlation"} and \code{"covariance"}, the estimates are based
on the sample complex covariance and use complex correlation \link[complex](ccor) and complex
covariance \link[complex](ccov) respectively. Note that the function does not calculate values for
lag 0. Also, the function will automatically remove NAs. Finally, function does not have
\code{demean} parameter, because \code{ccov()} and \code{ccor()} do that automatically.

The generic function plot has a method for objects of class "cacf".

The lag is returned and plotted in units of time, and not numbers of observations.

There is a print method for objects of class "cacf".
}
\examples{

# Generate random complex variables
x <- complex(real=rnorm(100,10,10), imaginary=rnorm(100,10,10))

# Calculate CACF
cacf(x)

}
\seealso{
\link[stats]{acf}, \link[complex]{ccor}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{univar}
