% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAR.R
\name{CAR}
\alias{CAR}
\title{Complex Autoregression}
\usage{
CAR(y, order = NULL, h = 13, holdout = TRUE, silent = TRUE,
  ic = c("AIC", "AICc", "BIC", "BICc"), restrict = TRUE)
}
\arguments{
\item{y}{The pair of the time series (the matrix) in columns.}

\item{order}{The order p of the model. If \code{NULL}, then it is selected
automatically based on the selected information criterion.}

\item{h}{Length of forecasting horizon.}

\item{holdout}{If \code{TRUE}, holdout sample of size \code{h} is taken from
the end of the data.
nothing happens.}

\item{silent}{If \code{TRUE}, then the plot is produced. Otherwise nothing happens.}

\item{ic}{The information criterion used in the model selection procedure.}

\item{restrict}{If \code{FALSE}, then VAR(p) is constructed instead of CAR(p).}
}
\value{
Object of class "vsmooth" is returned. It contains the following list of
values:
\itemize{
\item \code{model} - The name of the fitted model;
\item \code{timeElapsed} - The time elapsed for the construction of the model;
\item \code{coefficients} - The vector of all the estimated coefficients;
\item \code{nParam} - The number of estimated parameters;
\item \code{y} - The matrix with the original data;
\item \code{fitted} - The matrix of the fitted values;
\item \code{holdout} - The matrix with the holdout values (if \code{holdout=TRUE} in
the estimation);
\item \code{residuals} - The matrix of the residuals of the model;
\item \code{forecast} - The matrix of point forecasts;
\item \code{Sigma} - The covariance matrix of the errors (estimated with the correction
for the number of degrees of freedom);
\item \code{ICs} - The values of the information criteria for the selected model;
\item \code{logLik} - The log-likelihood function;
\item \code{lossValue} - The value of the loss function;
\item \code{loss} - The type of the used loss function;
\item \code{accuracy} - the values of the error measures. Currently not available.
\item \code{ICsAll} - In case of the order selection, the values of information
criteria for the tested models.
}
}
\description{
Function constructs Complex Autoregression model as a restricted VAR for a
provided pair time series.
}
\details{
The CAR(p) model can be written as a VAR(p) model with the following
restrictions on the proportionality matrices:
\deqn{
A_j = \begin{matrix} a_{j,1} & -a_{j,2} \\ a_{j,2} & a_{j,1} \end{matrix}
}{
A_j =
[ a_{j,1} | -a_{j,2} ]
[ a_{j,2} |  a_{j,1} ]
}

For some more information about the model and its implementation, see the
vignette: \code{vignette("CAR","Complex")}
}
\examples{

Y <- ts(cbind(rnorm(100,100,10),rnorm(100,75,8)),frequency=4)

# The simplest model applied to the data with the default values
CAR(Y,h=10,holdout=TRUE)


}
\seealso{
\code{\link[smooth]{ves}}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{models}
\keyword{regression}
\keyword{ts}
\keyword{univar}
