% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvar.R
\name{cvar}
\alias{cvar}
\alias{ccov}
\alias{ccor}
\alias{ccov2ccor}
\title{Complex Correlation, Variance and Covariance (Matrices)}
\usage{
cvar(x, ...)

ccov(x, y, ...)

ccor(x, y, ...)

ccov2ccor(V)
}
\arguments{
\item{x}{vector or matrix of complex variables. If it is matrix then the
variable \code{y} is ignored.}

\item{...}{parameters passed to \code{mean()} functions. For example, this can be
\code{na.rm=TRUE} to remove missing values or \code{trim} to define the trimming
in the mean (see \link[stats]{mean}).}

\item{y}{second vector to calculate covariance or correlations with.}

\item{V}{complex covariance matrix.}
}
\value{
A scalar or a matrix with resulting complex variables.

#' @references \itemize{
\item Svetunkov, S. (2022) Complex Autoregressions. In Press.
}
}
\description{
Functions return complex variance, covariance and correlation based on the
provided complex vector/matrix \code{x}.
}
\details{
Functions calculate complex statistics for complex variables. Only the parametric
correlation is supported by the function. If \code{x} is matrix, then \code{y}
is ignored.
}
\examples{

# Generate random complex variables
x <- complex(real=rnorm(100,10,10), imaginary=rnorm(100,10,10))
y <- complex(real=rnorm(100,10,10), imaginary=rnorm(100,10,10))

# Create a matrix of complex variables
z <- cbind(x,y)

# Calculate measures
cvar(x)
cvar(z)
ccor(x,y)
ccor(z)

}
\seealso{
\code{\link[stats]{cor}}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{univar}
