% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf.R
\name{pacf}
\alias{pacf}
\alias{print.pacf}
\alias{plot.pacf}
\title{Pseudo Correlation Function Estimation}
\usage{
pacf(x, lag.max = NULL, type = c("covariance", "correlation"), plot = TRUE)

\method{print}{pacf}(x, ...)

\method{plot}{pacf}(x, which = c(1, 2), ask = length(which) > 1, ...)
}
\arguments{
\item{x}{vector of complex variables.}

\item{lag.max}{maximum number of lags. See \link[stats]{acf} for more details.}

\item{type}{character string giving the type of PACF to be computed. Allowed values
are "correlation" (the default) and "covariance". Will be partially matched.}

\item{plot}{logical. If \code{TRUE} (the default) the PACF is plotted on complex plane
and as two linear graphs for real and imaginary parts.}

\item{which}{Determines, which of the plots to produce. 1 is the plot of real
and imaginary parts. 2 is the plot of absolute value and the argument.}

\item{ask}{Determines, whether to ask before producing a new plot or not.}
}
\value{
An object of class "pacf", which is a list with the following elements:
\itemize{
\item \code{lag} A three dimensional array containing the lags at which the PACF is estimated.
\item \code{acf} An array with the same dimensions as lag containing the estimated PACF.
\item \code{type} The type of correlation (same as the type argument).
\item \code{n.used} The number of observations in the time series.
\item \code{series} The name of the series x.
}

#' @references \itemize{
\item Svetunkov, S. (2022) Complex Autoregressions. In Press.
}
}
\description{
The function computes (and by default plots) estimates of the Pseudo Covariance
or Pseudo Correlation function.
}
\details{
For \code{type="correlation"} and \code{"covariance"}, the estimates are based
on the sample pseudo covariance and use pseudo correlation \link[complex]{pcor} and complex
covariance \link[complex]{pcov} respectively. Note that the function does not calculate values for
lag 0. Also, the function will automatically remove NAs. Finally, function does not have
\code{demean} parameter (as, for example, is done in \link[stats]{acf}), because \code{pcov()}
and \code{pcor()} do that automatically.

The generic function plot has a method for objects of class "pacf".

The lag is returned and plotted in units of time, and not numbers of observations.

There is a print and plot methods for objects of class "pacf".
}
\examples{

# Generate random complex variables
x <- complex(real=rnorm(100,10,10), imaginary=rnorm(100,10,10))

# Calculate PACF
pacf(x)

}
\seealso{
\link[stats]{acf}, \link[complex]{pcor}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{univar}
