% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf.R
\name{pcf}
\alias{pcf}
\title{Pseudo Correlation Function Estimation}
\usage{
pcf(x, lag.max = NULL, type = c("correlation", "covariance"), plot = TRUE)
}
\arguments{
\item{x}{vector of complex variables.}

\item{lag.max}{maximum number of lags. See \link[stats]{acf} for more details.}

\item{type}{character string giving the type of pcf to be computed. Allowed values
are "correlation" (the default) and "covariance". Will be partially matched.}

\item{plot}{logical. If \code{TRUE} (the default) the pcf is plotted on complex plane
and as two linear graphs for real and imaginary parts.}
}
\value{
An object of class "pcf", which is a list with the following elements:
\itemize{
\item \code{lag} A three dimensional array containing the lags at which the pcf is estimated.
\item \code{acf} An array with the same dimensions as lag containing the estimated pcf.
\item \code{type} The type of correlation (same as the type argument).
\item \code{n.used} The number of observations in the time series.
\item \code{series} The name of the series x.
}

#' @references \itemize{
\item Svetunkov, S. (2022) Complex Autoregressions. In Press.
}
}
\description{
The function computes (and by default plots) estimates of the Pseudo Covariance
or Pseudo Correlation function.
}
\details{
For \code{type="correlation"} and \code{"covariance"}, the estimates are based
on the sample pseudo covariance and use pseudo correlation \link[complex]{pcor} and complex
covariance \link[complex]{pcov} respectively. Note that the function does not calculate values for
lag 0. Also, the function will automatically remove NAs. Finally, function does not have
\code{demean} parameter (as, for example, is done in \link[stats]{acf}), because \code{pcov()}
and \code{pcor()} do that automatically.

The generic function plot has a method for objects of class "pcf".

The lag is returned and plotted in units of time, and not numbers of observations.

There is a print method for objects of class "pcf".
}
\examples{

# Generate random complex variables
x <- complex(real=rnorm(100,10,10), imaginary=rnorm(100,10,10))

# Calculate PCF
pcf(x)

}
\seealso{
\link[stats]{acf}, \link[complex]{pcor}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{univar}
