% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvar.R
\name{pvar}
\alias{pvar}
\alias{pcov}
\alias{pcor}
\alias{pcov2pcor}
\title{Pseudo Correlation, Variance and Covariance (Matrices)}
\usage{
pvar(x, ...)

pcov(x, y, ...)

pcor(x, y, ...)

pcov2pcor(V)
}
\arguments{
\item{x}{vector or matrix of complex variables. If it is matrix then the
variable \code{y} is ignored.}

\item{...}{parameters passed to \code{mean()} functions. For example, this can be
\code{na.rm=TRUE} to remove missing values or \code{trim} to define the trimming
in the mean (see \link[base]{mean}).}

\item{y}{second vector to calculate pseudo-covariance or pseudo-correlations with.}

\item{V}{complex pseudo-covariance matrix.}
}
\value{
A scalar or a matrix with resulting complex variables.

#' @references \itemize{
\item Svetunkov, S. (2022) Complex Autoregressions. In Press.
}
}
\description{
Functions return complex pseudo-variance, pseudo-covariance and pseudo-correlation based on the
provided complex vector/matrix \code{x}.
}
\details{
Functions calculate pseudo statistics for complex variables. Only the parametric
pseudo-correlation is supported by the function. If \code{x} is matrix, then \code{y}
is ignored.
}
\examples{

# Generate random complex variables
x <- complex(real=rnorm(100,10,10), imaginary=rnorm(100,10,10))
y <- complex(real=rnorm(100,10,10), imaginary=rnorm(100,10,10))

# Create a matrix of complex variables
z <- cbind(x,y)

# Calculate measures
pvar(x)
pvar(z)
pcor(x,y)
pcor(z)

}
\seealso{
\code{\link[stats]{cor}}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{univar}
