% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacf.R
\name{cacf}
\alias{cacf}
\alias{pcacf}
\alias{print.cacf}
\alias{plot.cacf}
\title{Complex Correlation Function Estimation}
\usage{
cacf(x, lag.max = NULL, type = c("correlation", "covariance"), plot = TRUE)

pcacf(x, lag.max = NULL, plot = TRUE)

\method{print}{cacf}(x, ...)

\method{plot}{cacf}(x, which = c(1, 2), ask = length(which) > 1, ...)
}
\arguments{
\item{x}{vector of complex variables.}

\item{lag.max}{maximum number of lags. See \link[stats]{acf} for more details.}

\item{type}{character string giving the type of CACF to be computed. Allowed values
are "correlation" (the default) and "covariance". Will be partially matched.}

\item{plot}{logical. If \code{TRUE} (the default) the CACF is plotted on complex plane
and as two linear graphs for real and imaginary parts.}

\item{...}{Parameter needed for consistency with print() and plot() generics.}

\item{which}{Determines, which of the plots to produce. 1 is the plot of real
and imaginary parts. 2 is the plot of absolute value and the argument.}

\item{ask}{Determines, whether to ask before producing a new plot or not.}
}
\value{
An object of class "cacf", which is a list with the following elements:
\itemize{
\item \code{lag} A three dimensional array containing the lags at which the CACF is estimated.
\item \code{acf} An array with the same dimensions as lag containing the estimated CACF.
\item \code{type} The type of correlation (same as the type argument).
\item \code{n.used} The number of observations in the time series.
\item \code{series} The name of the series x.
}

#' @references \itemize{
\item Svetunkov, S. (2022) Complex Autoregressions. In Press.
}
}
\description{
The functions compute (and by default plot) estimates of the Complex Autocovariance,
or Complex Autocorrelation, or Partial Complex Autocorrelation functions.
}
\details{
For \code{type="correlation"} and \code{"covariance"}, the estimates are based
on the sample pseudo covariance and use pseudo correlation \link[complex]{ccor} and complex
covariance \link[complex]{ccov} respectively. Note that the function does not calculate values for
lag 0. Also, the function will automatically remove NAs. Finally, function does not have
\code{demean} parameter (as, for example, is done in \link[stats]{acf}), because \code{ccov()}
and \code{ccor()} do that automatically.

\code{pcacf()} produces the partial complex ACF based on complex regression model of variable
on its lags.

The generic function plot has a method for objects of class "cacf".

The lag is returned and plotted in units of time, and not numbers of observations.

There is a print and plot methods for objects of class "cacf".
}
\examples{

# Generate random complex variables
x <- complex(real=rnorm(100,10,10), imaginary=rnorm(100,10,10))

# Calculate CACF
cacf(x)

}
\seealso{
\link[stats]{acf}, \link[complex]{ccor}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{univar}
